Update README.md
Student Management System is software which is helpful for students as well as the school
authorities. In the current system all the activities are done manually. Its time saving and
scalable. Our Student Management System deals with the various activities related to the
students
In the software we can register as a user and user has two types student and administrator.
Administrator has the power to add new user and can edit the students details entered. A
admin can add students record ,attendance status with department wise. All students can
search his/her basics details and attendance status with there respective roll numbers.